feat: Add setupIntegrations
