+++ src/Integration/Integration.php
@@ -5,7 +5,7 @@ namespace Sentry\Integration;
 interface Integration
 {
     /**
-     * Initializes the Integration once
+     * Initializes the Integration once.
      */
     public function setupOnce(): void;
 }
